% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_sme_functions.R
\name{CalculateSlurryConversion}
\alias{CalculateSlurryConversion}
\title{Function calculates slurry concentrations, densities and solids ratios}
\usage{
CalculateSlurryConversion(
  API_Key,
  url,
  p_liquid = NULL,
  p_solid = NULL,
  p_slurry = NULL,
  ratio_solid = NULL
)
}
\arguments{
\item{url}{API url}

\item{p_liquid}{Density of the liquid}

\item{p_solid}{Density of the solid}

\item{p_slurry}{Density of the solid}

\item{APIKey}{API key provided}

\item{ratio_slurry}{Solid percent ratio}
}
\description{
Placeholder description
}
\section{More details}{
\itemize{
\item Bullet 1
\item Bullet 2
\item etc
}
}

\examples{

\dontrun{

Calculate the density of the liquid
CalculateSlurryConversion(API_Key = API_Key,
                           url = url,
                           # p_solid = 3900,
                           p_slurry = 1341,
                           ratio_solid = 0.25,
                           p_liquid = NULL)

}

}
