% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_sme_functions.R
\name{CalculateTwoProductRecovery}
\alias{CalculateTwoProductRecovery}
\title{Function calculates Feed, concentrate, tail and recovery}
\usage{
CalculateTwoProductRecovery(
  API_Key,
  url,
  feed = NULL,
  concentrate = NULL,
  tail = NULL,
  recovery = NULL
)
}
\arguments{
\item{API_Key}{API key provided}

\item{url}{API url}

\item{feed}{Value of the feed}

\item{concentrate}{Value of the concentrate}

\item{tail}{Value of the tail}

\item{recovery}{Value of the recovery}
}
\description{
Placeholder
}
\section{More details}{
-Bullet 1
-Bullet2
}

\examples{

\dontrun{

CalculateTwoProductRecovery(API_Key = API_Key,
                                  url = url,
                                 feed = 0.7,
                          concentrate = 25,
                                 # tail= 0.02,
                             recovery = 0.9722)


}

}
