% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClarofyCS.R
\docType{package}
\name{ClarofyCS}
\alias{ClarofyCS}
\alias{ClarofyCS-package}
\title{ClarofyCS: A package to use the Clarofy API.}
\description{
This package provides functionality for making various calls to the API
}
\author{
Matt Rosinski \email{matt.rosinski@interlate.com}
}
