% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyEndSession}
\alias{ClarofyEndSession}
\title{Finish the current session}
\usage{
ClarofyEndSession(url, id)
}
\arguments{
\item{url}{API url}

\item{id}{Session id}
}
\description{
This function closed the current session
}
\details{
\itemize{
\item Closes the current session
}
}
\examples{

\dontrun{

ClarofyEndSession(url = url, id = sessionID)

}

}
