% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-in-one-functions.R
\name{ClarofyFinalOptimiseStart}
\alias{ClarofyFinalOptimiseStart}
\title{Start an all in one optimisation process}
\usage{
ClarofyFinalOptimiseStart(
  url,
  APIKey,
  train,
  test,
  Dependent_Variable,
  Independent_Variables,
  Process_Control_Variables,
  Upstream_Forecasting_Variables,
  number_clusters,
  maximise_dependent_variable = TRUE,
  monitor_status = TRUE
)
}
\arguments{
\item{url}{API url}

\item{APIKey}{API key provided}

\item{train}{training data set}

\item{test}{test data set}

\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{Independent_Variables}{regression variables that are used to predict dependent variable}

\item{Process_Control_Variables}{a subset of independent variables that you want to optimise}

\item{Upstream_Forecasting_Variables}{a subset of independent variables which you can't easily control}

\item{number_clusters}{number of clusters to search for in upstream forecasting variables}

\item{maximise_dependent_variable}{logical to indicate whether to maximise or minimise the dependent variable}

\item{monitor_status}{logical to indicate whether session is to wait for the process to complete or return id for later retrieval}
}
\description{
Start the all in one optimisation workflow
}
\section{Optimisation Workflow}{
\itemize{
\item Returns the optimisation job identifier
\item Setting monitor_status = TRUE means your current session process will
be in a held state until the job is completed.
\item Setting monitor_status = FALSE means you can continue to use your current
session without waiting for the optimisation to complete.  You will be able
to use the returned jobId to fetch the optimisation results at a later time.
}
}

\examples{

\dontrun{

all_in_one_jobId <-
ClarofyFinalOptimiseStart(
                 train = train,
                 test = holdout,
                 Dependent_Variable = Dependent_Variable,
                 Independent_Variables = Independent_Variables,
                 Process_Control_Variables = Process_Control_Variables,
                 Upstream_Forecasting_Variables = Upstream_Forecasting_Variables,
                 number_clusters = 2,
                 maximise_dependent_variable = FALSE,
                 monitor_status = TRUE,
                 url = url,
                 APIKey = APIKey)

}

}
