% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-in-one-functions.R
\name{ClarofyFinalOptimiseStatus}
\alias{ClarofyFinalOptimiseStatus}
\title{Check status of all-in-one optimisation workflow}
\usage{
ClarofyFinalOptimiseStatus(url, id, jobId)
}
\arguments{
\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}

\item{jobId}{job id for temporary processes run during a session}
}
\description{
Returns status of all-in-one optimisation workflow
}
\section{Optimisation Workflow}{
\itemize{
\item Returns optimised predictions for the dependent variable using the optimal
process control variables
}
}

\examples{

\dontrun{

all_in_one_jobStatus <-
ClarofyFinalOptimiseStatus(url = url, id = sessionID, jobId = all_in_one_jobId)

}

}
