% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyFit}
\alias{ClarofyFit}
\title{Train a model on the session training data}
\usage{
ClarofyFit(Dependent_Variable, Independent_Variables, url, id)
}
\arguments{
\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{Independent_Variables}{regression variables that are used to predict dependent variable}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Train a model on the session training data
Returns returns details for the trained model
}
\section{Optimisation Workflow}{
\itemize{
\item Train a model for optimisation of your process control variables
}
}

\examples{

\dontrun{

Model_Trainer_Output <- ClarofyFit(Dependent_Variable = Dependent_Variable,
                                  Independent_Variables = Independent_Variables,
                                  url = url,
                                  id = sessionID)

}

}
