% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-in-one-functions.R
\name{ClarofyGetProductivityGroups}
\alias{ClarofyGetProductivityGroups}
\title{Get the productivity groups}
\usage{
ClarofyGetProductivityGroups(url, id)
}
\arguments{
\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Returns the productivity groups from the all-in-one optimisation workflow
}
\section{Optimisation Workflow}{
\itemize{
\item Returns the producitivity groups from the all-in-one optimisation workflow
}
}

\examples{

\dontrun{

productivityGroups <- ClarofyGetProductivityGroups(url = url, id = sessionID)

}

}
