% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyGroup}
\alias{ClarofyGroup}
\title{Generate productivity groups using the training data set}
\usage{
ClarofyGroup(
  Dependent_Variable,
  Upstream_Forecasting_Variables,
  url,
  id,
  number_clusters = 5
)
}
\arguments{
\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{Upstream_Forecasting_Variables}{a subset of independent variables which you can't easily control}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}

\item{number_clusters}{number of clusters to search for in upstream forecasting variables}
}
\description{
Productivity groups are used to generate optimised process control variable
recommendations for common operating scenarios.  A productivity group can be
thought of as a pattern of upstream forecasting variables that may be reflective
of different feed types or blends.
}
\section{Optimisation Workflow}{
\itemize{
\item Generated productivity groups for use in optimisation
}
}

\examples{

\dontrun{

ProductivityGroups <- ClarofyGroup(Dependent_Variable = Dependent_Variable,
Upstream_Forecasting_Variables = Upstream_Forecasting_Variables,
url = url, id = sessionID, number_clusters = 10)

}

}
