% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyOptimisedPredictions}
\alias{ClarofyOptimisedPredictions}
\title{Collect optimised predictions for the dependent variable if using the optimum values}
\usage{
ClarofyOptimisedPredictions(url, id)
}
\arguments{
\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Collect optimal results from job
}
\section{Optimisation Workflow}{
\itemize{
\item Returns optimised predictions for the dependent variable using the optimal
process control variables
}
}

\examples{

\dontrun{

OptimisedPredictions <- ClarofyOptimisedPredictions(url = url, id = sessionID)

}

}
