% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyOptimiseJobStatus}
\alias{ClarofyOptimiseJobStatus}
\title{Check on status of conditional optimisation job}
\usage{
ClarofyOptimiseJobStatus(url, id, jobId)
}
\arguments{
\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}

\item{jobId}{job id for temporary processes run during a session}
}
\description{
Returns the status of the optimisation job
}
\section{Optimisation Workflow}{
\itemize{
\item Returns status of optimisation job
}
}

\examples{

\dontrun{

ClarofyOptimiseJobStatus(url = url, id = sessionID, jobId = jobId)

}

}
