% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyOptimiseStartJob}
\alias{ClarofyOptimiseStartJob}
\title{Start optimisation job using a single productivity group or time weighted aggregate}
\usage{
ClarofyOptimiseStartJob(
  Dependent_Variable,
  Process_Control_Variables,
  Upstream_Forecasting_Variables,
  upstream_values,
  maximise_dependent_variable = TRUE,
  monitor_status = TRUE,
  url,
  id
)
}
\arguments{
\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{Process_Control_Variables}{a subset of independent variables that you want to optimise}

\item{Upstream_Forecasting_Variables}{a subset of independent variables which you can't easily control}

\item{upstream_values}{the upstream variables from a selected productivity group or aggregate}

\item{maximise_dependent_variable}{logical to indicate whether to maximise or minimise the dependent variable}

\item{monitor_status}{logical to indicate whether session is to wait for the process to complete or return id for later retrieval}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Commence an optimisation job.
}
\section{Optimisation Workflow}{
\itemize{
\item Returns the jobId for later retrieval of the optimal values
}
}

\examples{

\dontrun{

jobId <- ClarofyOptimiseStartJob(Dependent_Variable = Dependent_Variable,
                                 Upstream_Forecasting_Variables = Upstream_Forecasting_Variables,
                                 Process_Control_Variables = Process_Control_Variables,
                                 upstream_values = upstream_values,
                                 maximise_dependent_variable = FALSE,
                                 monitor_status = TRUE,
                                 url = url,
                                 id = sessionID)

}

}
