% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyPlotOptimisationRun}
\alias{ClarofyPlotOptimisationRun}
\title{Plot optimisation results}
\usage{
ClarofyPlotOptimisationRun(Dependent_Variable, ncol, url, id)
}
\arguments{
\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{ncol}{Number of columns on the plot}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
This function plots the range of optimisation results
}
\section{Optimisation Workflow}{
\itemize{
\item Plots the optimisation results
}
}

\examples{

\dontrun{

ClarofyPlotOptimisationRun(
    Dependent_Variable = Dependent_Variable,
    url = url,
    id = sessionID)

}

}
