% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ClarofyPredict}
\alias{ClarofyPredict}
\title{Predict the dependent variable using the built model}
\usage{
ClarofyPredict(Dependent_Variable, url, id, df)
}
\arguments{
\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}

\item{df}{dataframe used to make new predictions}
}
\description{
Predict the dependent variable
}
\section{Optimisation Workflow}{
\itemize{
\item Make predictions with trained model
}
}

\examples{

\dontrun{

Data_Predictions_train <- ClarofyPredict(Dependent_Variable = Dependent_Variable,
url = url, id = sessionID, data = train)

}

}
