% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{CleanseAllTestingData}
\alias{CleanseAllTestingData}
\title{Cleanse session training data using density cleaning operation}
\usage{
CleanseAllTestingData(Dependent_Variable, Independent_Variables, url, id)
}
\arguments{
\item{Dependent_Variable}{dependent variable for which you want to predict or optimise}

\item{Independent_Variables}{regression variables that are used to predict dependent variable}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Cleanse training data using density cleansing operation
Returns cleaned data with outliers replaced with NAs.
Does not alter training data.
Use \link{LoadTrainingDataIntoSession} if wanting to train the model on this cleansed data.
NB: Training data must not contain NA values.
}
\section{Optimisation Workflow}{
\itemize{
\item Custom Clarofy cleaning algorithm for simultaneous cleaning of all variables.
\item This function cleansing the uploaded training data but does not change it.
\item The returned cleansed dataset can be inspected, compared and loaded as new training data if desired.
\item The best practice is to train models on cleaned data and test on uncleaned data.
}
}

\examples{

\dontrun{

Data_Cleaned <- CleanseAllTestingData(
   Dependent_Variable = Dependent_Variable,
   Independent_Variables = Independent_Variables,
   url = url, id = sessionID)

}
}
