% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{CreateSession}
\alias{CreateSession}
\title{Call Clarofy API endpoints}
\usage{
CreateSession(APIKey, url = "http://127.0.0.1:8000")
}
\arguments{
\item{APIKey}{API key provided}

\item{url}{API url}
}
\description{
Creates a new session to use the Clarofy API
}
\section{All In One Optimisation Workflow}{
\itemize{
\item Creates New API Session
}
}

\examples{

\dontrun{

url <- "SUBSTITUTE-API-URL"
APIKEY <- "YOUR-API-KEY"
sessionID <- CreateSession(APIKey = APIKey, url = url)

}
}
