% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{GetSummaryOfSessionData}
\alias{GetSummaryOfSessionData}
\title{Get loaded data back from session}
\usage{
GetSummaryOfSessionData(id, url)
}
\arguments{
\item{id}{Session id created by \link{CreateSession}}

\item{url}{API url}
}
\description{
Get data back from a Clarofy API session
}
\section{Optimisation Workflow}{
\itemize{
\item Get raw data back from session
}
}

\examples{

\dontrun{

Summary <- GetSummaryOfSessionData(url  = url, id   = sessionID)

}

}
