% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{ImportData}
\alias{ImportData}
\title{Load testing and demo data}
\usage{
ImportData(APIKey, url)
}
\arguments{
\item{APIKey}{API key provided}

\item{url}{API url}
}
\description{
Imports data for testing and demonstration purposes
}
\details{
\itemize{
\item Loads data for testing
}
}
\examples{

\dontrun{

Data <- ImportData(APIKey = APIKey, url = url)

}
}
