% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iron_flotation_data}
\alias{iron_flotation_data}
\title{Contains sample iron flotation time series data for optimisation}
\format{
A data.frame with 24 columns and 2539 rows
\describe{
\item{XIronConcentrate}{Primary Dependent variable - Iron concentration in flotation product that we want to maximise}
\item{XSilicaConcentrate}{Secondary Dependent variable - Silica concentration in flotation product that we want to minimise}
\item{AminaFlow}{Flotation regent dose rate as a process control variable}
\item{StarchFlow}{Flotation regent dose rate as a process control variable}
\item{TimeStamp}{Time formated column}
\item{Upstream Forecasting Variables}{The variables that impact on the dependent variables but which may not be process control variables}
}
}
\source{
Data are from a public Kaggle data set
}
\usage{
iron_flotation_data
}
\description{
Contains sample iron flotation time series data for optimisation
}
\examples{
iron_flotation_data

}
\keyword{datasets}
