% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{LoadDataFrameIntoSession}
\alias{LoadDataFrameIntoSession}
\title{Load data into session}
\usage{
LoadDataFrameIntoSession(df, url, id)
}
\arguments{
\item{df}{Dataframe to be loaded into session}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Load data into the Clarofy API session
}
\section{Optimisation Workflow}{
\itemize{
\item Load data into session
}
}

\examples{

\dontrun{

loadStatus <- LoadDataFrameIntoSession(
df      = Data,
url     = url,
id      = sessionID)

}

}
