% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{LoadTestingDataIntoSession}
\alias{LoadTestingDataIntoSession}
\title{Load testing data into session}
\usage{
LoadTestingDataIntoSession(test, url, id)
}
\arguments{
\item{test}{Model testing data}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Load testing data into Clarofy API session
Returns status of load
}
\section{Optimisation Workflow}{
\itemize{
\item Get raw data back from session
}
}

\examples{

\dontrun{

test_load_status <- LoadTestingDataIntoSession(test = holdout, url = url, id = sessionID)

}

}
