% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarofy_functions.R
\name{LoadTrainingDataIntoSession}
\alias{LoadTrainingDataIntoSession}
\title{Load training data into session}
\usage{
LoadTrainingDataIntoSession(train, url, id)
}
\arguments{
\item{train}{Model training data}

\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Load training data into Clarofy API session
}
\section{Optimisation Workflow}{
\itemize{
\item Get raw data back from session
}
}

\examples{

\dontrun{

train_load_status <- LoadTrainingDataIntoSession(train = train, url = url, id = sessionID)

}

}
