% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-in-one-functions.R
\name{ClarofyGetSensitivityAnalysis}
\alias{ClarofyGetSensitivityAnalysis}
\title{Get the sensitivity analysis from the all-in-one optimisation workflow}
\usage{
ClarofyGetSensitivityAnalysis(url, id)
}
\arguments{
\item{url}{API url}

\item{id}{Session id created by \link{CreateSession}}
}
\description{
Returns the sensitivity estimates from the all-in-one optimisation workflow
}
\section{Optimisation Workflow}{
\itemize{
\item Returns the sensitivity estimates from the all-in-one optimisation workflow
}
}

\examples{

\dontrun{

sensitivityAnalysis <-
ClarofyGetSensitivityAnalysis(url = url, id = sessionID)

}

}
